acl: Derive Clone on signature types
Replace the manual Copy/Clone impls with an automatic Clone. Maybe
some of these should actually be Copy, but the calling code generally
calls `.clone()` so prefer that for now.

No obvious change to benchmarks, although I'm seeing ~5% noise on
my machine so if there's a small effect it wouldn't be visible.
However, none of the custom implementations do anything I wouldn't
expect to see from the derived impl.