Use copied and cloned iter methods
Replace explicit map() calls with built-in copied() and cloned()
methods from the std library which do the same thing. Addresses
a clippy warning with rust 1.78.0.

I think there's some inconsistency between whether we clone or
copy; this doesn't try to resolve that, but it might be a place
to look for performance improvements.