github actions: Execute tests in release mode
Split the 'Run Tests' step into a parallel job and build with
the `--release` flag. Performance on intensive operations is
not good in the default debug builds, so the extra time building
with optimizations enabled pays off in faster test execution.

Currently it takes around 12 minutes to get ci feedback on changes.