bulletproofs: Remove inactive cfg_attr
The rest of the code assumes the `std` library, and no such feature
is defined in Cargo.toml, so this has no effect. Furthermore,
I don't think the `error` annotation has any effect without an
import of the `thiserror` crate or similar. Which would be fine,
but isn't currently part of the build. The comment is sufficient
in the meantime.