bulletproofs: clean up add_vec
Rewrite `utils::add_vec` to use iterators. This should generate
better code since it's easier to eliminate bounds checks.

Make the function private and replace the comment with an actual
assertion that the argument lengths match. This is only used
locally (so far) so it doesn't need to be accessible outside
the module.

Add some minimal test coverage.