Update to digest and sha3 v0.10
Migrate crates which were using the older v0.9 releases to
the current stable releases of these crates for easier
maintenance and integration with other software.

The `_dirty` variants are replaced with the default `finalize`
with a separate `_reset` variety to recover the old reset-for-
next use behavior. At least as far as I can tell.

Import the digest traits through sha3's re-export to ensure the
versions align.