demo: update to axum-server v0.7.1
Switch to the `tls-rustls-no-provider` feature added in v0.7.1.
Since v0.7.0 the `tls-rustls` feature uses rustls v0.23 which
require a per-process (or thread) `CryptoProvider` be set.
By default, rustls uses `aws_lc_rs` which requires `cmake` et al.
to build. Worse, `axum-server` doesn't re-export the relevant
`rustls` symbols, so one has to declare a direct `rustls` dependency
and keep it in sync to call the required initialization function.

Since this is a basic example, just disable to provider for simplicity.