demo: Generate a new self-signed certificate each time
In a public deployment one generally wants to use a cert signed
by a recognized authority, and loading such a cert from a file
is a common arrangement. Self-signed cert files are also simple
to generate, but this is an extra step that impedes trying out
the end-to-end examples. Hence, self-signed cert and private key
files were generated and added to the repository.

There is concern, however, that this cert will end up in public
use somewhere, which is insecure since the private key is published
in the repository. Likewise, it makes for extra audit work that
the key published here is not already being used elsewhere. For
these reasons, it's best not to include key material, even for
testing purposes.

Instead, generate an ephemeral cert and key in memory and use that.

This does make it more difficult to pass the cert to the client.